s3-unpack/pyproject.toml
[tool.poetry]
name = "s3-unpack"
version = "0.1.0"
description = "Container for clamav scanning objectstore objects."
authors = ["Per Buer <per.buer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyClamd = "^0.4.0"
azure-storage-blob = "^12.8.1"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
python-dotenv = "^0.13.0"
pytest = "^6.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
